requirements.txt
# Web framework
flask>=3.0.0

# CORS support for Flask
flask-cors>=4.0.0

# Claude API client
anthropic>=0.39.0

# Environment variable management
python-dotenv>=1.0.0

# HTTP requests
requests>=2.31.0
